requirements.txt
altgraph==0.17
APScheduler==3.7.0
beautiful-date==2.0.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
future==0.18.2
google-api-core==1.26.3
google-api-python-client==1.8.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
html2text==2020.1.16
httplib2==0.19.1
idna==2.10
importlib-metadata==4.0.1
MouseInfo==0.1.3
oauthlib==3.1.0
packaging==20.9
pefile==2019.4.18
Pillow==8.2.0
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.1.4
PyScreeze==0.1.27
python-dateutil==2.8.1
python-dotenv==0.17.1
PyTweening==1.0.3
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
zipp==3.4.1
